requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==5.4.1
    # via streamlit
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.8.2
    # via streamlit
cachetools==5.5.2
    # via streamlit
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via streamlit
colorama==0.4.6
    # via click
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dotenv==0.9.9
    # via comp-escalavel-a2 (pyproject.toml)
faker==37.4.0
    # via
    #   comp-escalavel-a2 (pyproject.toml)
    #   faker-commerce
faker-commerce==1.0.4
    # via comp-escalavel-a2 (pyproject.toml)
fonttools==4.58.4
    # via matplotlib
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
idna==3.10
    # via requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.24.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
kafka-python==2.2.13
    # via comp-escalavel-a2 (pyproject.toml)
kiwisolver==1.4.8
    # via matplotlib
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.3
    # via comp-escalavel-a2 (pyproject.toml)
narwhals==1.43.1
    # via
    #   altair
    #   plotly
numpy==2.3.1
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   pydeck
    #   streamlit
packaging==25.0
    # via
    #   altair
    #   matplotlib
    #   plotly
    #   streamlit
pandas==2.3.0
    # via
    #   comp-escalavel-a2 (pyproject.toml)
    #   streamlit
pillow==11.2.1
    # via
    #   matplotlib
    #   streamlit
plotly==6.1.2
    # via comp-escalavel-a2 (pyproject.toml)
protobuf==6.31.1
    # via streamlit
psycopg2-binary==2.9.9    # via comp-escalavel-a2 (pyproject.toml)
py4j==0.10.9.9
    # via pyspark
pyarrow==20.0.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
pyparsing==3.2.3
    # via matplotlib
pyspark==4.0.0
    # via comp-escalavel-a2 (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.1.0
    # via dotenv
pytz==2025.2
    # via pandas
redis==6.2.0
    # via comp-escalavel-a2 (pyproject.toml)
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via streamlit
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.46.0
    # via
    #   comp-escalavel-a2 (pyproject.toml)
    #   streamlit-autorefresh
streamlit-autorefresh==1.0.1
    # via comp-escalavel-a2 (pyproject.toml)
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.1
    # via streamlit
typing-extensions==4.14.0
    # via
    #   altair
    #   referencing
    #   streamlit
tzdata==2025.2
    # via
    #   faker
    #   pandas
urllib3==2.5.0
    # via requests
watchdog==6.0.0
    # via streamlit